Initial commit of haxe
Summary:
Requirement of various games and apps, such as ldtk.

OCaml and co. are required for the parser, such dependencies are built from source.
As for the `String::ShellQuote` Perl module, the thing doesn't display a clear license from the sources and is rock ancient.

Depends on D11420 and D11421.

Signed-off-by: Martin Reboredo <yakoyoku@gmail.com>

Test Plan: Built (and patched) Friday Night Funkin and played it.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D11422